requirements.txt
autopep8
flake8
Flask
#Flask-MonitoringDashboard
google-cloud-logging
pytest
pytest-cov
python-amazon-paapi
python-dotenv
airtable-python-wrapper
